feat: serve legacy void.cat uploads
